requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.4.2
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2-binary==2.9.7
redis==5.0.0
celery==5.3.4
requests==2.31.0
aiohttp==3.8.5

# Data Processing
numpy==1.26.0
pandas==2.1.1
scipy==1.11.3
scikit-learn==1.3.1
statsmodels==0.14.0

# Geospatial Processing
geopandas==0.14.0
folium==0.14.0
geopy==2.4.0
osmnx==1.6.0
networkx==3.1

# Machine Learning
tensorflow==2.14.0
prophet==1.1.4

# Monitoring and Logging
prometheus-client==0.17.1
python-json-logger==2.0.7
sentry-sdk==1.31.0

# Testing
pytest==7.4.2
pytest-cov==4.1.1
pytest-mock==3.11.1
responses==0.23.3

# Development
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
pre-commit==3.4.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# API Clients
# These would be replaced with your actual inventory/transport/weather API clients
inventoryapi-client==1.0.0
transportapi-client==1.0.0
openweathermap-client==1.0.0
